+++ board-frontend/html/scripts/protocol/client/ClientMessageType.js
@@ -4,6 +4,7 @@ const ClientMessageType = {
     CREATE_DOCUMENT:2,
     HANDSHAKE:3,
     DELETE_DOCUMENT:4,
-    UPDATE_DOCUMENT:5
+    UPDATE_DOCUMENT:5,
+    GET_INVITE:6,
 };
 export default ClientMessageType;